requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg[binary]
alembic
httpx